Removed failing new standardisation (does not really work and causes Asserts with MSVC6). Corrected population decrease. Removed update option from advanced settings. Revisited the Disband code to prevent Asserts. BOOL/sint32 to bool conversions and other small cleanups.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@721 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/ui/aui_common/aui_surface.cpp
@@ -309,7 +309,7 @@ AUI_ERRCODE aui_Surface::GetDC(HDC * hdc)
 	bih.biWidth = m_width;
 	bih.biHeight = -m_height;
 	bih.biPlanes = 1;
-	bih.biBitCount = GetDeviceCaps( m_hdc, BITSPIXEL );
+	bih.biBitCount = static_cast<WORD>(GetDeviceCaps(m_hdc, BITSPIXEL));
 	bih.biCompression = BI_RGB;
 
 	SetDIBits(
@@ -340,7 +340,7 @@ AUI_ERRCODE aui_Surface::ReleaseDC(HDC hdc)
 	bih.biWidth = m_width;
 	bih.biHeight = -m_height;
 	bih.biPlanes = 1;
-	bih.biBitCount = GetDeviceCaps( m_hdc, BITSPIXEL );
+	bih.biBitCount = static_cast<WORD>(GetDeviceCaps(m_hdc, BITSPIXEL));
 	bih.biCompression = BI_RGB;
 
 	GetDIBits(
